RankScore test: Introduce variant UID for dataset comparision, low bound check
It was observed that in comparing reference and test set,
variants rank scores got mixed up due to identical
VCF.ID values, causing the RankScore test to fail.

As a solution, introduce a new UID based on VCF fields:
 * POS
 * ID
 * ALT
 * REF

so that there can be no ambiguity in variant sorting in
the dataframe.

Furthermore, allow RankScores to differ in between
reference set and test set, due to newly introduced
RankScore capping behavior in Genmod.
See more information in patch itself.

Signed-off-by: Tor Björgen <tor.bjorgen@scilifelab.se>