VRS: Drop test dataset shuffle and increase batch size
Since the model is not dependant on batch size
in the model validation step,
and the n_steps/epoch is covering all of the
validation data, it's OK to remove
shuffle and increase batch size for performance
reasons.

Since the n epochs during validation fit step anyhow covers
all of the data per epoch.

Signed-off-by: Tor Björgen <tor.bjorgen@scilifelab.se>